+++ demos/modules/demo_tables.js
@@ -1,5 +1,5 @@
 /**
- * NAME: tables.js
+ * NAME: demo_tables.js
  * AUTH: Brent Ely (https://github.com/gitbrent/)
  * DESC: Common test/demo slides for all library features
  * DEPS: Used by various demos (./demos/browser, ./demos/node, etc.)
